Added maxcfpeqmaxhist option.
With this option, if the user wants to plot a histogram and CFP
together, it is really easy. There was a way to use `--maxhistone` in
conjunction with `--normcfp`, but in some conditions that is not
enough. For example if you want to plot two histograms and only the
CFP of one of them.